requirements.txt
pyyaml
jinja2
markdown-it-py